asset style fixed

+++ backend/assetmanagement/models.py
@@ -93,7 +93,7 @@ class Asset(models.Model):
     status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='active')
     purchase_date = models.DateField()
     purchase_price = models.DecimalField(max_digits=12, decimal_places=2)
-    current_value = models.DecimalField(max_digits=12, decimal_places=2)
+    current_value = models.IntegerField()
     serial_number = models.CharField(max_length=100, blank=True)
     asset_code = models.CharField(max_length=100, blank=True)
     brand = models.CharField(max_length=100, blank=True)
